Added default

+++ latest/inquisiv-startup.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 
 # Checkout svn
-svn checkout $2 $1/web2py/ --username $3 --password $4 --non-interactive --trust-server-cert
-cp $1/web2py/parameters_8000.py $1/web2py/parameters_80.py
+if ! [ -f $1/web2py/applications/inquisiv/models/_config.py ]
+then
+	svn checkout $2 $1/web2py/ --username $3 --password $4 --non-interactive --trust-server-cert
+	cp $1/web2py/parameters_8000.py $1/web2py/parameters_80.py
+fi
 
 # Start emperor
 echo Start uwsgi-emperor
